remove useless conversion of type

+++ pkg/tracer/span.go
@@ -150,7 +150,7 @@ func ZapFieldToOpentracing(zapField zapcore.Field) opentracinglog.Field {
 	case zapcore.Float64Type:
 		return opentracinglog.Float64(zapField.Key, math.Float64frombits(uint64(zapField.Integer)))
 	case zapcore.Int64Type:
-		return opentracinglog.Int64(zapField.Key, int64(zapField.Integer))
+		return opentracinglog.Int64(zapField.Key, zapField.Integer)
 	case zapcore.Int32Type:
 		return opentracinglog.Int32(zapField.Key, int32(zapField.Integer))
 	case zapcore.StringType:
